Refactor players indexes
