+++ src/containers/Players/handlers/index.ts
@@ -1,0 +1,1 @@
+export * from './players.handler';
