+++ src/containers/Players/handlers/players.handler.ts
@@ -1,16 +1,17 @@
 import { rest } from 'msw';
 
-import { yahooPlayersMock } from '../mocks/players.mocks';
+import { yahooPlayersMock } from '../mocks/';
 
 const { ENDPOINT } = process.env;
 
-const handler = rest.get(`${ENDPOINT}/players`, (req, res, ctx) => {
-	return res(
-		ctx.json({
-			players: yahooPlayersMock,
-			statusFilters: ['N/A', 'IL10', 'DTD'],
-		})
-	);
-});
-
-export default handler;
+export const playersHandler = rest.get(
+	`${ENDPOINT}/players`,
+	(req, res, ctx) => {
+		return res(
+			ctx.json({
+				players: yahooPlayersMock,
+				statusFilters: ['N/A', 'IL10', 'DTD'],
+			})
+		);
+	}
+);
