+++ src/containers/Players/index.ts
@@ -1,37 +1,7 @@
-import { getPlayers } from './api/getPlayers';
-import { PlayerStatusMap } from './data/playerStatusMap';
-import playersHandler from './handlers/players.handler';
-import useGetPlayersQueryResponse from './hooks/useGetPlayersQueryResponse';
-import { draftKingsPlayersMock, yahooPlayersMock } from './mocks/players.mocks';
-import { IDraftKingsPlayer } from './models/IDraftKingsPlayer';
-import { IPlayer } from './models/IPlayer';
-import { TPlayerStatus, TPlayerStatusColor } from './models/IPlayerStatus';
-import { IYahooPlayer } from './models/IYahooPlayer';
-import { mapDraftKingsPlayers, mapYahooPlayers } from './services/mapPlayers';
-
-// Data
-export { PlayerStatusMap };
-
-// Api
-export { getPlayers };
-
-// Mocks
-export { draftKingsPlayersMock, yahooPlayersMock };
-
-// Handler
-export { playersHandler };
-
-// Services
-export { mapDraftKingsPlayers, mapYahooPlayers };
-
-// Models
-export type {
-	IDraftKingsPlayer,
-	IPlayer,
-	IYahooPlayer,
-	TPlayerStatus,
-	TPlayerStatusColor,
-};
-
-// Hooks
-export { useGetPlayersQueryResponse };
+export * from './api';
+export * from './data';
+export * from './handlers';
+export * from './hooks';
+export * from './mocks';
+export * from './models';
+export * from './services';
