+++ src/containers/Players/mocks/index.ts
@@ -1,0 +1,1 @@
+export * from './players.mocks';
